WishlistItem refactored to accept customerId instead of customer object.  Create, get and delete mapping is functional.

+++ src/main/java/org/launchcode/books2borrow/controllers/WishlistController.java
@@ -1,12 +1,16 @@
 package org.launchcode.books2borrow.controllers;
 
 
+import org.launchcode.books2borrow.data.BookRepository;
 import org.launchcode.books2borrow.data.CustomerRepository;
 import org.launchcode.books2borrow.data.WishlistItemRepository;
+import org.launchcode.books2borrow.models.Book;
+import org.launchcode.books2borrow.models.Customer;
 import org.launchcode.books2borrow.models.WishlistItem;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
@@ -23,28 +27,79 @@ public class WishlistController {
     @Autowired
     private CustomerRepository customerRepository;
 
-    @GetMapping("{userId}")
-    public ResponseEntity<?> wishlist(@PathVariable Integer userId) {
-        List<WishlistItem> customersWishlist = customerRepository.findById(Long.valueOf(userId)).get().getWishlist();
+    @Autowired
+    private BookRepository bookRepository;
+
+    @RequestMapping("/all")
+    public ResponseEntity<?> wishlist(Authentication authentication) {
+        int userId;
+        List<Customer> customers = customerRepository.findByEmail(authentication.getName());
+        if (!customers.isEmpty()) {
+            userId = customers.get(0).getId();
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        List<WishlistItem> customersWishlist = wishlistItemRepository.findByCustomerId(userId);
+
         return new ResponseEntity<>(customersWishlist, HttpStatus.OK);
     }
 
-    @PostMapping("addToWishlist")
-    public ResponseEntity<String> addToWishlist(@RequestBody WishlistItem wishlistItem) {
+    @PostMapping("/addToWishlist")
+    public ResponseEntity<?> addToWishlist(Authentication authentication, @RequestBody WishlistItem partialWishlistItem) {
+    //Map<String, String> wishlistMap) {
+        int customerId;
+        List<Customer> customers = customerRepository.findByEmail(authentication.getName());
+        if (!customers.isEmpty()) {
+            customerId = customers.get(0).getId();
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        String bookKey = partialWishlistItem.getBookKey();
+        String title = partialWishlistItem.getTitle();
+        int bookCover = partialWishlistItem.getBookCover();
+
+//        String bookKey = wishlistMap.bookKey;
+//        String title = wishlistMap.title;
+//        int bookCover = Integer.valueOf(wishlistMap.bookCover);
+        boolean isAvailable = false;
+        List<Book> matchingBooks = bookRepository.findByBookKey(bookKey);
+            for(int i = 0; i < matchingBooks.size(); i++) {
+                if (matchingBooks.get(i).isAvailable()) {
+                    isAvailable = true;
+                    break;
+                }
+            };
+        WishlistItem wishlistItem = new WishlistItem(customerId, bookKey, title, bookCover, isAvailable);
         wishlistItemRepository.save(wishlistItem);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<?> deleteWishItem(@PathVariable Integer id) {
+    public ResponseEntity<?> deleteWishItem(Authentication authentication, @PathVariable int id) {
+        int userId;
+        List<Customer> customers = customerRepository.findByEmail(authentication.getName());
+        if (!customers.isEmpty()) {
+            userId = customers.get(0).getId();
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         Optional<WishlistItem> wishlistOptional = wishlistItemRepository.findById(id);
-        if (wishlistOptional.isPresent()) {
+        if (wishlistOptional.isPresent() && (wishlistOptional.get().getCustomerId() == userId)) {
             wishlistItemRepository.deleteById(id);
             return new ResponseEntity<>(HttpStatus.OK);
         } else {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
     }
+
+
+//    Customer user;
+//    List<Customer> customers = customerRepository.findByEmail(authentication.getName());
+//        if (!customers.isEmpty()) {
+//        user = customers.get(0);
+//    } else {
+//        user = null;
+//    }
 
 
 }
